+++ src/app/not-found.tsx
@@ -1,5 +1,5 @@
 import NotFound from '@/components/pages/NotFound'
 
-export default function NotFoundGlobal() {
-  return <NotFound />
+export default function GlobalNotFound() {
+    return <NotFound />
 }
